refactor: move the skills-tab builders into a shared module
技能タブの行データ（組込・基本・カスタム技能と技能ポイント）の組み立てを
context/character.ts から context/skills.ts へ出す。共鳴者と人間NPCは
CharacterLikeDataModel を共有しており、行の形も同じになるため、アクター型を
その交差型に広げてどちらのシートからも呼べるようにする。
